+++ checkout/internal/usecase/delete_from_cart.go
@@ -4,6 +4,6 @@ import (
 	"context"
 )
 
-func (u *useCase) DeleteFromCart(ctx context.Context, user int64, sku uint32, count uint16) error {
+func (u *useCase) DeleteFromCart(ctx context.Context, user int64, sku uint32, count uint32) error {
 	return nil
 }
